Cleaned up the Code Nicely
