+++ data/school_admin.txt
@@ -1,0 +1,30 @@
+The Principal of Sainik School Kazhakootam is Col.Dhirendra Kumar here is more about him:
+
+Col Dhirendra Kumar assumed appointment as Principal of Sainik School Kazhakootam w.e.f. 25 Aug 2021.
+
+
+
+He is a P.G. & M Phil holder from Jawaharlal Nehru University, New Delhi. He is also B.Ed,NET & JRF qualified. He has been posted at various Army Establishments, Training Centres and devotedly worked in enhancing educational qualification of Army Personnel.
+
+
+
+Col Dhirendra Kumar for his exceptional contribution to the organization and exhibiting dedication and devotion towards his duties has been awarded with VCoAS Commendation Card in 2012 and again GOC in C Commendation Card in 2016.
+
+
+
+Col Dhirendra Kumar has experience of serving in Sainik School, he had served as Vice Principal in Sainik School Balachadi for 3 years. During his tenure, there was a marked improvement in academic and NDA results.
+
+
+The Vice-Principal of Saink School Kazhakootam is Wg Cdr. M Rajkumar,here is more about him:
+Wg Cdr Rajkumar Marudhavanan, the 22nd Vice Principal of Sainik School Kazhakootam, was commissioned in the Education Branch of Indian Air Force on 16 Dec 2007 from Air Force Academy, Dundigal, Hyderabad. The Officer is a Graduate in Economics and a Post Graduate in Personnel Management. He is also a Post Graduate Diploma holder in Yoga. He has presented and published two research papers on “Corporate Yoga” and “Industries’ Expectations from the University” in the international Seminar conducted by APS University at Rewa in 2018.
+
+Wg Cdr Rajkumar Marudhavanan, the 22nd Vice Principal of Sainik School Kazhakootam, was commissioned in the Education Branch of Indian Air Force on 16 Dec 2007 from Air Force Academy, Dundigal, Hyderabad. The Officer is a Graduate in Economics and a Post Graduate in Personnel Management. He is also a Post Graduate Diploma holder in Yoga. He has presented and published two research papers on “Corporate Yoga” and “Industries’ Expectations from the University” in the international Seminar conducted by APS University at Rewa in 2018.
+
+
+
+The administrative Officer of Sainik School Kazhakootam is Cdr. Sarin Pulloor Madathil 
+
+The senior master of Sainik School Kazhakootam is Mr. Mahadevan Nair G
+
+
+
