+++ requirements.txt
@@ -1,0 +1,84 @@
+aiohttp==3.8.5
+aiosignal==1.3.1
+altair==5.0.1
+async-timeout==4.0.2
+attrs==23.1.0
+beautifulsoup4==4.12.2
+blinker==1.6.2
+cachetools==5.3.1
+certifi==2023.7.22
+charset-normalizer==3.2.0
+click==8.1.6
+colorama==0.4.6
+comtypes==1.2.0
+dataclasses-json==0.5.13
+decorator==5.1.1
+frozenlist==1.4.0
+fsspec==2023.6.0
+gitdb==4.0.10
+GitPython==3.1.32
+greenlet==2.0.2
+idna==3.4
+importlib-metadata==6.8.0
+Jinja2==3.1.2
+jsonschema==4.18.4
+jsonschema-specifications==2023.7.1
+keyboard==0.13.5
+langchain==0.0.240
+langsmith==0.0.14
+llama-index==0.7.11.post1
+markdown-it-py==3.0.0
+MarkupSafe==2.1.3
+marshmallow==3.20.1
+mdurl==0.1.2
+multidict==6.0.4
+mypy-extensions==1.0.0
+nest-asyncio==1.5.6
+numexpr==2.8.4
+numpy==1.25.1
+openai==0.27.8
+openapi-schema-pydantic==1.2.4
+packaging==23.1
+pandas==2.0.3
+Pillow==9.5.0
+protobuf==4.23.4
+pyarrow==12.0.1
+PyAudio==0.2.13
+pydantic==1.10.11
+pydeck==0.8.0
+Pygments==2.15.1
+Pympler==1.0.1
+pypiwin32==223
+python-dateutil==2.8.2
+pyttsx3==2.90
+pytz==2023.3
+pytz-deprecation-shim==0.1.0.post0
+pywin32==306
+PyYAML==6.0.1
+referencing==0.30.0
+regex==2023.6.3
+requests==2.31.0
+rich==13.4.2
+rpds-py==0.9.2
+six==1.16.0
+smmap==5.0.0
+soupsieve==2.4.1
+SpeechRecognition==3.10.0
+SQLAlchemy==2.0.19
+streamlit==1.25.0
+streamlit-chat==0.1.1
+tenacity==8.2.2
+tiktoken==0.4.0
+toml==0.10.2
+toolz==0.12.0
+tornado==6.3.2
+tqdm==4.65.0
+typing-inspect==0.9.0
+typing_extensions==4.7.1
+tzdata==2023.3
+tzlocal==4.3.1
+urllib3==1.26.16
+validators==0.20.0
+watchdog==3.0.0
+yarl==1.9.2
+zipp==3.16.2
