+++ storage/graph_store.json
@@ -1,0 +1,1 @@
+{"graph_dict": {}}
