+++ storage/index.json
@@ -1,0 +1,1 @@
+{"index_store/data": {"f852a358-cb76-431d-b43f-df8b60e06b28": {"__type__": "vector_store", "__data__": "{\"index_id\": \"f852a358-cb76-431d-b43f-df8b60e06b28\", \"summary\": null, \"nodes_dict\": {\"55c6b43b-3edb-46fa-aaba-9c508367901c\": \"55c6b43b-3edb-46fa-aaba-9c508367901c\", \"73f102a4-401c-4f63-a492-5939420f712b\": \"73f102a4-401c-4f63-a492-5939420f712b\"}, \"doc_id_dict\": {}, \"embeddings_dict\": {}}"}}}
