+++ storage/index_store.json
@@ -1,0 +1,1 @@
+{"index_store/data": {"7bbb35c7-615d-4658-862d-f3b10b9f8f0a": {"__type__": "vector_store", "__data__": "{\"index_id\": \"7bbb35c7-615d-4658-862d-f3b10b9f8f0a\", \"summary\": null, \"nodes_dict\": {\"ad905bd2-46f3-4249-8a30-7c7c62fc5edd\": \"ad905bd2-46f3-4249-8a30-7c7c62fc5edd\", \"82d6ff67-158c-45fb-acad-5c74b8b57801\": \"82d6ff67-158c-45fb-acad-5c74b8b57801\", \"a4d16494-418f-44e2-aa07-1f530dce7539\": \"a4d16494-418f-44e2-aa07-1f530dce7539\", \"7e1f06f4-f378-48cf-bfaa-c3a2c0c9a1bc\": \"7e1f06f4-f378-48cf-bfaa-c3a2c0c9a1bc\", \"7cc3d0b9-5d65-4e0f-9637-04bc2d631468\": \"7cc3d0b9-5d65-4e0f-9637-04bc2d631468\"}, \"doc_id_dict\": {}, \"embeddings_dict\": {}}"}}}
